feat: use strangelove style guide
