+++ .vscode/settings.json
@@ -14,15 +14,5 @@
   "css.validate": false,
   "editor.formatOnSave": true,
   "eslint.packageManager": "yarn",
-  "eslint.workingDirectories": [
-    {
-      "!cwd": false,
-      "pattern": "packages/*"
-    },
-    {
-      "!cwd": false,
-      "directory": "."
-    }
-  ],
   "npm.packageManager": "yarn"
 }
