+++ package.json
@@ -1,9 +1,6 @@
 {
   "name": "juno-tools",
   "version": "1.3.0",
-  "workspaces": [
-    "packages/*"
-  ],
   "scripts": {
     "build": "next build",
     "dev": "next dev",
@@ -51,28 +48,19 @@
     "zustand": "^3"
   },
   "devDependencies": {
+    "@strangelovelabs/style-guide": "latest",
     "@types/node": "^14",
     "@types/react": "^18",
     "@types/react-datetime-picker": "^3",
-    "@types/react-dom": "^17.0.11",
+    "@types/react-dom": "^18",
     "autoprefixer": "^10",
+    "eslint": "^8",
     "husky": "^7",
     "lint-staged": "^12",
     "postcss": "^8",
+    "prettier": "^2",
     "tailwindcss": "^3",
     "typescript": "^4"
-  },
-  "eslintConfig": {
-    "extends": [
-      "@juno-tools/eslint-config"
-    ],
-    "ignorePatterns": [
-      ".next",
-      "node_modules",
-      "out",
-      "public"
-    ],
-    "root": true
   },
   "lint-staged": {
     "*.{json,md}": [
@@ -82,6 +70,12 @@
       "eslint --fix"
     ]
   },
-  "prettier": "@juno-tools/prettier-config",
+  "prettier": {
+    "endOfLine": "auto",
+    "printWidth": 120,
+    "semi": false,
+    "singleQuote": true,
+    "trailingComma": "all"
+  },
   "private": true
 }
