--- packages/prettier-config/index.js
@@ -1,12 +1,0 @@
-// @ts-check
-
-/** @type {import("prettier").Config} */
-const prettierConfig = {
-  endOfLine: 'auto',
-  printWidth: 120,
-  semi: false,
-  singleQuote: true,
-  trailingComma: 'all',
-}
-
-module.exports = prettierConfig
