--- packages/prettier-config/package.json
@@ -1,26 +1,0 @@
-{
-  "name": "@juno-tools/prettier-config",
-  "version": "0.0.0",
-  "dependencies": {
-    "prettier": "^2"
-  },
-  "devDependencies": {
-    "@types/prettier": "^2"
-  },
-  "eslintConfig": {
-    "extends": [
-      "@juno-tools/eslint-config"
-    ],
-    "ignorePatterns": [
-      "node_modules"
-    ],
-    "root": true
-  },
-  "lint-staged": {
-    "*.{js,json,md}": [
-      "prettier --write"
-    ]
-  },
-  "prettier": "./index.js",
-  "license": "MIT"
-}
